added anolytics functions
