+++ caliban_toolbox/figure_eight_functions_test.py
@@ -30,6 +30,7 @@ import requests_mock
 import urllib
 import pathlib
 import zipfile
+import tempfile
 
 import numpy as np
 import pandas as pd
@@ -57,60 +58,6 @@ class FakeResponse(object):
         self.status_code = status_code
 
 
-def test_get_latest_log_file(tmp_path):
-    upload_logs = ['stage_0_upload_log.csv', 'stage_3_upload_log.csv',
-                   'stage_8_upload_log.csv']
-
-    for log in upload_logs:
-        Path(os.path.join(tmp_path, log)).touch()
-
-    latest_file = figure_eight_functions.get_latest_log_file(tmp_path)
-    assert latest_file == 'stage_8_upload_log.csv'
-
-
-def test_create_job_urls(tmp_path):
-    npz_files = ['test_0.npz', 'test_1.npz', 'test_2.npz']
-
-    for npz in npz_files:
-        Path(os.path.join(tmp_path, npz)).touch()
-
-    aws_folder = 'aws_main_folder/aws_sub_folder'
-    stage = 'test_stage'
-    pixel_only, label_only, rgb_mode = True, False, True
-
-    output_lists = figure_eight_functions.create_job_urls(crop_dir=tmp_path,
-                                                          aws_folder=aws_folder,
-                                                          stage=stage,
-                                                          pixel_only=pixel_only,
-                                                          label_only=label_only,
-                                                          rgb_mode=rgb_mode)
-
-    npz_paths, npz_keys, url_paths, npzs_to_upload = output_lists
-
-    assert len(npz_paths) == len(npz_keys) == len(url_paths) == len(npzs_to_upload) == 3
-
-    # TODO: Figure out how we're going to validate inputs
-    # with tempfile.TemporaryDirectory() as temp_dir:
-    #     # NPZ name with spaces leads to bad url
-    #     npz_files = ['bad file name.npz']
-    #
-    #     for npz in npz_files:
-    #         Path(os.path.join(temp_dir, npz)).touch()
-    #
-    #     aws_folder = 'aws_main_folder/aws_sub_folder'
-    #     stage = 'test_stage'
-    #     pixel_only, label_only, rgb_mode = True, False, True
-    #
-    #     with pytest.raises(ValueError):
-    #
-    #         output_lists = figure_eight_functions.create_job_urls(crop_dir=temp_dir,
-    #                                                               aws_folder=aws_folder,
-    #                                                               stage=stage,
-    #                                                               pixel_only=pixel_only,
-    #                                                               label_only=label_only,
-    #                                                               rgb_mode=rgb_mode)
-
-
 def test_copy_job():
     with requests_mock.Mocker() as m:
 
@@ -128,6 +75,27 @@ def test_copy_job():
         new_job_id = figure_eight_functions.copy_job(job_id=test_job_id, key=test_appen_key)
 
         assert new_job_id == return_id
+
+
+def test_create_upload_log():
+    filepaths = ['file_path_' + str(x) for x in range(10)]
+    filenames = ['file_name_' + str(x) for x in range(10)]
+    stage = 'all_the_world_is_a_stage'
+    aws_folder = 'aws_folder_path'
+    job_id = '007'
+
+    with tempfile.TemporaryDirectory() as temp_dir:
+        log_name = 'test_log.csv'
+        figure_eight_functions.create_upload_log(base_dir=temp_dir, stage=stage,
+                                                 aws_folder=aws_folder,
+                                                 filenames=filenames, filepaths=filepaths,
+                                                 job_id=job_id,
+                                                 pixel_only=False, rgb_mode=True, label_only=True,
+                                                 log_name=log_name)
+
+        log_file = pd.read_csv(os.path.join(temp_dir, 'logs', log_name))
+
+        assert np.all(log_file['filename'] == filenames)
 
 
 def test_upload_log_file(mocker):
