--- caliban_toolbox/log_file_test.py
@@ -1,51 +1,0 @@
-# Copyright 2016-2020 The Van Valen Lab at the California Institute of
-# Technology (Caltech), with support from the Paul Allen Family Foundation,
-# Google, & National Institutes of Health (NIH) under Grant U24CA224309-01.
-# All rights reserved.
-#
-# Licensed under a modified Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.github.com/vanvalenlab/caliban-toolbox/LICENSE
-#
-# The Work provided may be used for non-commercial academic purposes only.
-# For any other use of the Work, including commercial use, please contact:
-# vanvalenlab@gmail.com
-#
-# Neither the name of Caltech nor the names of its contributors may be used
-# to endorse or promote products derived from this software without specific
-# prior written permission.
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-import tempfile
-import os
-
-import pandas as pd
-import numpy as np
-
-from caliban_toolbox.log_file import create_upload_log
-
-
-def test_create_upload_log():
-    filepaths = ['file_path_' + str(x) for x in range(10)]
-    filenames = ['file_name_' + str(x) for x in range(10)]
-    stage = 'all_the_world_is_a_stage'
-    aws_folder = 'aws_folder_path'
-    job_id = '007'
-
-    with tempfile.TemporaryDirectory() as temp_dir:
-        log_name = 'test_log.csv'
-        create_upload_log(base_dir=temp_dir, stage=stage, aws_folder=aws_folder,
-                          filenames=filenames, filepaths=filepaths, job_id=job_id,
-                          pixel_only=False, rgb_mode=True, label_only=True,
-                          log_name=log_name)
-
-        log_file = pd.read_csv(os.path.join(temp_dir, 'logs', log_name))
-
-        assert np.all(log_file['filename'] == filenames)
